requirements.txt
# Complete Email Extraction System Dependencies

# Core automation and web scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# HTTP client
httpx>=0.25.0
requests>=2.31.0

# DNS validation
dnspython>=2.4.0

# Environment variables
python-dotenv>=1.0.0

# AI/LLM
groq>=0.4.0

# Search
duckduckgo-search>=3.9.0

# Data processing
pandas>=2.0.0
